moved the file to the right location
