+++ csharp/ROS.Domain/ROS.Domain/ROS.Domain.csproj
@@ -114,6 +114,7 @@
     <Compile Include="Models\User.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\AddressContactService.cs" />
+    <Compile Include="Services\AdminService.cs" />
     <Compile Include="Services\BoatService.cs" />
     <Compile Include="Services\ClubService.cs" />
     <Compile Include="Services\EntryServie.cs" />
